+++ crates/rust-client/src/builder.rs
@@ -9,6 +9,7 @@ use miden_tx::auth::TransactionAuthenticator;
 use rand::Rng;
 
 use crate::keystore::FilesystemKeyStore;
+use crate::note_transport::NoteTransportClient;
 use crate::rpc::NodeRpcClient;
 use crate::store::Store;
 use crate::{Client, ClientError, DebugMode};
@@ -59,6 +60,8 @@ pub struct ClientBuilder<AUTH> {
     /// Maximum number of blocks the client can be behind the network for transactions and account
     /// proofs to be considered valid.
     max_block_number_delta: Option<u32>,
+    /// An optional custom note transport client.
+    note_transport_api: Option<Arc<dyn NoteTransportClient>>,
 }
 
 impl<AUTH> Default for ClientBuilder<AUTH> {
@@ -72,6 +75,7 @@ impl<AUTH> Default for ClientBuilder<AUTH> {
             in_debug_mode: DebugMode::Disabled,
             tx_graceful_blocks: Some(TX_GRACEFUL_BLOCKS),
             max_block_number_delta: None,
+            note_transport_api: None,
         }
     }
 }
@@ -157,6 +161,13 @@ where
         self
     }
 
+    /// Sets a custom note transport client directly.
+    #[must_use]
+    pub fn note_transport(mut self, client: Arc<dyn NoteTransportClient>) -> Self {
+        self.note_transport_api = Some(client);
+        self
+    }
+
     /// Build and return the `Client`.
     ///
     /// # Errors
@@ -220,6 +231,7 @@ where
             .expect("Default executor's options should always be valid"),
             self.tx_graceful_blocks,
             self.max_block_number_delta,
+            self.note_transport_api,
         )
         .await
     }