+++ crates/testing/miden-client-tests/src/tests.rs
@@ -33,6 +33,7 @@ use miden_client::testing::common::{
     setup_wallet_and_faucet,
 };
 use miden_client::testing::mock::{MockClient, MockRpcApi};
+use miden_client::testing::note_transport::{MockNoteTransportApi, MockNoteTransportNode};
 use miden_client::transaction::{
     DiscardCause,
     PaymentNoteDescription,
@@ -42,14 +43,15 @@ use miden_client::transaction::{
     TransactionRequestError,
     TransactionStatus,
 };
+use miden_client::utils::RwLock;
 use miden_client::{ClientError, DebugMode};
 use miden_client_sqlite_store::SqliteStore;
 use miden_lib::account::auth::AuthRpoFalcon512;
 use miden_lib::account::faucets::BasicFungibleFaucet;
 use miden_lib::account::interface::AccountInterfaceError;
 use miden_lib::account::wallets::BasicWallet;
-use miden_lib::note::utils;
 use miden_lib::note::well_known_note::WellKnownNote;
+use miden_lib::note::{create_p2id_note, utils};
 use miden_lib::testing::mock_account::MockAccountExt;
 use miden_lib::testing::note::NoteBuilder;
 use miden_lib::transaction::TransactionKernel;
@@ -2248,4 +2250,74 @@ async fn insert_new_fungible_faucet(
 
     client.add_account(&account, false).await?;
     Ok(account)
+}
+
+pub async fn create_test_client_transport(
+    mock_ntnode: Arc<RwLock<MockNoteTransportNode>>,
+) -> (MockClient<FilesystemKeyStore<StdRng>>, FilesystemKeyStore<StdRng>) {
+    let (builder, _, keystore) = create_test_client_builder().await;
+    let transport_client = MockNoteTransportApi::new(mock_ntnode);
+    let builder_w_transport = builder.note_transport(Arc::new(transport_client));
+
+    let mut client = builder_w_transport.build().await.unwrap();
+    client.ensure_genesis_in_place().await.unwrap();
+
+    (client, keystore)
+}
+
+pub async fn create_test_user_transport(
+    mock_ntnode: Arc<RwLock<MockNoteTransportNode>>,
+) -> (MockClient<FilesystemKeyStore<StdRng>>, Account) {
+    let (mut client, keystore) = Box::pin(create_test_client_transport(mock_ntnode.clone())).await;
+    let account = insert_new_wallet(&mut client, AccountStorageMode::Private, &keystore)
+        .await
+        .unwrap();
+    (client, account)
+}
+
+#[tokio::test]
+async fn transport_basic() {
+    // Setup entities
+    let mock_ntnode = Arc::new(RwLock::new(MockNoteTransportNode::new()));
+    let (mut sender, sender_account) = create_test_user_transport(mock_ntnode.clone()).await;
+    let (mut recipient, recipient_account) = create_test_user_transport(mock_ntnode.clone()).await;
+    let recipient_address =
+        Address::from(AccountIdAddress::new(recipient_account.id(), AddressInterface::BasicWallet));
+    let (mut observer, _observer_account) = create_test_user_transport(mock_ntnode.clone()).await;
+
+    // Create note
+    let note = create_p2id_note(
+        sender_account.id(),
+        recipient_account.id(),
+        vec![],
+        NoteType::Private,
+        Felt::default(),
+        sender.rng(),
+    )
+    .unwrap();
+
+    // Sync-state / fetch notes
+    // No notes before sending
+    recipient.sync_state().await.unwrap();
+    let notes = recipient.get_input_notes(NoteFilter::All).await.unwrap();
+    assert_eq!(notes.len(), 0);
+
+    // Send note
+    sender.send_private_note(note, &recipient_address).await.unwrap();
+
+    // Sync-state / fetch notes
+    // 1 note stored
+    recipient.sync_state().await.unwrap();
+    let notes = recipient.get_input_notes(NoteFilter::All).await.unwrap();
+    assert_eq!(notes.len(), 1);
+
+    // Sync again, should be only 1 note stored
+    recipient.sync_state().await.unwrap();
+    let notes = recipient.get_input_notes(NoteFilter::All).await.unwrap();
+    assert_eq!(notes.len(), 1);
+
+    // Third user shouldn't receive any note
+    observer.sync_state().await.unwrap();
+    let notes = observer.get_input_notes(NoteFilter::All).await.unwrap();
+    assert_eq!(notes.len(), 0);
 }